IS-2007: Refactor folder structure inside mote folder
